Add JSON

+++ ESAlignTool/test/OnLxplus/createAndSubmitJobsOnCMSDATA.py
@@ -101,14 +101,23 @@ BATCHDIR=${PWD}
 EOSPATH="EOS_PATH"
 
 #export SCRAM_ARCH=slc5_amd64_gcc462
-export SCRAM_ARCH=slc6_amd64_gcc481
+#export SCRAM_ARCH=slc6_amd64_gcc481
+export SCRAM_ARCH=slc6_amd64_gcc491
 cd MAIN_WORKDIR
 eval `scram runtime -sh`
 
 cp -v MAIN_WORKDIR/CMSSW_cfg.py $BATCHDIR/CMSSW_cfg.py
 cp -v MAIN_WORKDIR/inputMatrixElements_cfi.py $BATCHDIR/
-cp -v MAIN_WORKDIR/inputDB.db $BATCHDIR/
 cp -v DATASET_WORKDIR/input/inputFiles_JOB_NUMBER_cfi.py $BATCHDIR/inputFiles_cfi.py
+
+if [ -e MAIN_WORKDIR/inputDB.db ]
+then
+    cp -v MAIN_WORKDIR/inputDB.db $BATCHDIR/
+fi
+if [ -e MAIN_WORKDIR/myJSON.txt ]
+then
+    cp -v MAIN_WORKDIR/myJSON.txt $BATCHDIR/
+fi
 
 cd $BATCHDIR
 echo "Running CMSSW job"
@@ -143,6 +152,7 @@ def main():
   parser.add_option("-w", "--main_workdir", dest="main_workdir", action='store', help="Main working directory", metavar="MAIN_WORKDIR")
   parser.add_option("-d", "--dataset_list", dest="dataset_list", action='store', help="Text file containing a list of datasets to be processed", metavar="DATASET_LIST")
   parser.add_option("-D", "--DBFileInput",  dest="DBFileInput",  action='store', help="Input DB file", default='', metavar="DBFILEINPUT")
+  parser.add_option("-J", "--myJSONfile", dest="myJSONfile", action='store', help="JSON file for selecting event", metavar="MYJSONFILE")
   parser.add_option("-o", "--output_filename", dest="output_filename", action='store', default='AlignmentFile', help="Output ROOT filename (Default set to AlignmentFile)", metavar="OUTPUT_FILENAME")
   parser.add_option("-E", "--eos_path", dest="eos_path", action='store', help="EOS path to copy output files to (This parameter is optional)", metavar="EOS_PATH")
   parser.add_option('-m', '--match', dest="match", action='store', help='Only files containing the MATCH string in their names will be considered (This parameter is optional)', metavar='MATCH')
@@ -178,6 +188,8 @@ def main():
   # copy the CMSSW cfg file to the cfg_files_dir
   shutil.copyfile(cmssw_cfg,os.path.join(main_workdir,'CMSSW_cfg.py'))
 
+  if options.myJSONfile != '':
+    shutil.copyfile( options.myJSONfile, os.path.join(main_workdir,'myJSON.txt'))
   if options.DBFileInput != '':
     shutil.copyfile( options.DBFileInput, os.path.join(main_workdir,'inputDB.db'))
